Rearranging files for Receipt.js
